refactor(scripts): rename + tighten 2 utility Python scripts
Rename to snake_case so pylint's invalid-name (C0103) passes, then
fix the rest of the PYTHON_BLACK/FLAKE8/PYLINT/RUFF/ISORT findings:

- supermicro-updater.py → supermicro_updater.py
  - Add module + function docstrings; move OpenSSL.crypto / http.client
    imports to top of file (C0415); replace exit() with sys.exit()
    (R1722); narrow broad `except Exception` to requests.RequestException
    / KeyError / IndexError / ValueError / json.JSONDecodeError (W0718);
    .format() → lazy %-style logger args (C0209); cert_data.split(s)[0]
    → split(s, maxsplit=1)[0] (C0207); rename setLogger → set_logger
    (C0103); split main() into helpers to drop branch count (R0912).
- sync-exclusions.py → sync_exclusions.py
  - Drop unused `import sys` (F401); add docstrings; open() now passes
    encoding="utf-8" (W1514); narrow broad Exception to a tuple of
    urllib.error.URLError / TimeoutError / json.JSONDecodeError / OSError.

Reference updates that follow the renames:
- exclusion-sync kustomization.yaml + cronjob.yaml use the new filename.
- supermicro/CONTAINERIZED-README.md filename mentions updated; the doc
  was also prettier-reformatted in the same pass.
