+++ kubernetes/apps/infrastructure/certwarden/cert-deployment/supermicro/CONTAINERIZED-README.md
@@ -7,31 +7,32 @@ This deployment now uses a **pre-built container** for faster and more reliable 
 - **Container**: `ghcr.io/lukeevanstech/supermicro-ipmi-cert:latest`
 - **Script**: `certwarden-supermicro-deploy.sh`
 - **Benefits**:
-  - Faster execution (no runtime dependency installation)
-  - Consistent environment across deployments
-  - Reduced job execution time by ~30-60 seconds
-  - Better security (minimal container, no runtime downloads)
+    - Faster execution (no runtime dependency installation)
+    - Consistent environment across deployments
+    - Reduced job execution time by ~30-60 seconds
+    - Better security (minimal container, no runtime downloads)
 
 ## Files
 
 - **certwarden-supermicro-deploy.sh** - Active script (containerized approach)
 - **certwarden-supermicro-deploy.sh.OLD** - Backup of old script (runtime installation)
-- **supermicro-updater.py** - Python script (kept for reference, baked into container)
+- **supermicro_updater.py** - Python script (kept for reference, baked into container)
 
 ## How It Works
 
 1. Certwarden calls `certwarden-supermicro-deploy.sh` after certificate renewal
 2. Script creates a Kubernetes Job using the pre-built container
 3. Container has all dependencies pre-installed:
-   - kubectl v1.32.0
-   - Python 3.12 + requests + pyOpenSSL
-   - supermicro_ipmi_cert.py script
+    - kubectl v1.32.0
+    - Python 3.12 + requests + pyOpenSSL
+    - supermicro_ipmi_cert.py script
 4. Job deploys certificate via Redfish API
 5. Job auto-cleans up after 5 minutes
 
 ## Testing
 
 Test the deployment through the Certwarden UI:
+
 1. Go to Certwarden UI
 2. Select a Supermicro certificate
 3. Click "Renew" or trigger post-processing
@@ -43,24 +44,26 @@ Test the deployment through the Certwarden UI:
 If you need to rollback to the old runtime-installation method:
 
 1. Edit `kustomization.yaml`:
-   ```yaml
-   # Comment out containerized approach:
-   # - name: certwarden-supermicro-scripts
-   #   files:
-   #     - certwarden-supermicro-deploy.sh
 
-   # Uncomment old approach:
-   - name: certwarden-supermicro-scripts
-     files:
-       - supermicro-updater.py
-       - certwarden-supermicro-deploy.sh.OLD
-   ```
+    ```yaml
+    # Comment out containerized approach:
+    # - name: certwarden-supermicro-scripts
+    #   files:
+    #     - certwarden-supermicro-deploy.sh
+
+    # Uncomment old approach:
+    - name: certwarden-supermicro-scripts
+      files:
+          - supermicro_updater.py
+          - certwarden-supermicro-deploy.sh.OLD
+    ```
 
 2. Apply changes:
-   ```bash
-   kubectl delete configmap certwarden-supermicro-scripts -n infrastructure
-   kubectl apply -k .
-   ```
+
+    ```bash
+    kubectl delete configmap certwarden-supermicro-scripts -n infrastructure
+    kubectl apply -k .
+    ```
 
 3. Restart Certwarden to pick up the new ConfigMap
 
@@ -73,14 +76,17 @@ Updates are automatically built and pushed to GHCR when changes are committed to
 ## Troubleshooting
 
 **Job fails with ImagePullBackOff:**
+
 - Check container registry: `docker pull ghcr.io/lukeevanstech/supermicro-ipmi-cert:latest`
 - Verify GHCR is accessible from cluster
 
 **Job fails with "kubectl not found":**
+
 - Container should have kubectl pre-installed at `/usr/local/bin/kubectl`
 - Check container build logs in GitHub Actions
 
 **Certificate upload fails:**
+
 - Same troubleshooting as old method
 - Check IPMI credentials in ExternalSecret
 - Verify Redfish API is enabled on IPMI
