+++ kubernetes/apps/media/exclusion-sync/app/cronjob.yaml
@@ -25,7 +25,7 @@ spec:
               image: docker.io/python:3.14-alpine@sha256:dd4d2bd5b53d9b25a51da13addf2be586beebd5387e289e798e4083d94ca837a
               command:
                 - "python3"
-                - "/scripts/sync-exclusions.py"
+                - "/scripts/sync_exclusions.py"
               env:
                 - name: RADARR_URL
                   value: "https://radarr.codelooks.com"
